requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.3.2
requests==2.28.2
psutil==5.9.4
pydantic==1.10.7
python-dotenv==1.0.0
drf-yasg==1.21.5